+++ src/Job.h
@@ -7,15 +7,15 @@ using namespace std;
 class Job {
 private:
     int jobNumber;
-    int pageCount;
-    int pagesToPrint;
+    double pageCount;
+    double pagesToPrint;
     string userName;
 
-    int start_time;
+    double start_time;
     bool finished;
 
 public:
-    Job(int jobNumber, int pageCount, string userName): jobNumber(jobNumber), pageCount(pageCount), userName(userName){
+    Job(int jobNumber, double pageCount, string userName): jobNumber(jobNumber), pageCount(pageCount), userName(userName){
         finished = false;
         pagesToPrint = pageCount;
     }
